Move patch_adk from __init__.py to patchers.py for a cleaner namespace structure
